refactor : RequestBodyJsonGenerator utils 패키지로 이동
