+++ src/test/java/com/peer/missionpeerflow/MissionPeerFlowAnswerAPITests.java
@@ -1,6 +1,7 @@
 package com.peer.missionpeerflow;
 
 
+import com.peer.missionpeerflow.utils.RequestBodyJsonGenerator;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
